remove MockReportGen
