+++ inference_perf/reportgen/base.py
@@ -12,10 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import json
+import statistics
 from pydantic import BaseModel
-from abc import ABC, abstractmethod
-from typing import Any, List, Tuple
-from inference_perf.metrics import MetricsClient
+from typing import Any, List
+from inference_perf.metrics import MetricsClient, MetricsSummary
 
 class ReportFile():
     name: str
@@ -38,6 +38,7 @@ class ReportFile():
     def get_contents(self) -> dict[str, Any]:
         return self.contents.model_dump()
 
+
 class RequestMetric(BaseModel):
     stage_id: int
     prompt_tokens: int
@@ -45,16 +46,32 @@ class RequestMetric(BaseModel):
     time_per_request: float
 
 
-class ReportGenerator(ABC):
-    @abstractmethod
-    def __init__(self, metrics_client: MetricsClient, *args: Tuple[int, ...]) -> None:
-        self.metrics_client = metrics_client
-        pass
+class ReportGenerator():
+    def __init__(self, metrics_client: MetricsClient) -> None:
+        super().__init__(metrics_client = metrics_client)
+        self.metrics: List[RequestMetric] = []
 
-    @abstractmethod
     def collect_request_metrics(self, metric: RequestMetric) -> None:
-        raise NotImplementedError
+        self.metrics.append(metric)
 
-    @abstractmethod
     async def generate_reports(self) -> List[ReportFile]:
-        raise NotImplementedError
+        print("\n\nGenerating Report ..")
+        summary = self.metrics_client.collect_metrics_summary()
+        if summary is not None:
+            for field_name, value in summary:
+                print(f"{field_name}: {value}")
+
+        elif summary is None and len(self.metrics) > 0:
+            summary = MetricsSummary(
+                total_requests=len(self.metrics),
+                avg_prompt_tokens=statistics.mean([x.prompt_tokens for x in self.metrics]),
+                avg_output_tokens=statistics.mean([x.output_tokens for x in self.metrics]),
+                avg_time_per_request=statistics.mean([x.time_per_request for x in self.metrics]),
+            )
+            for field_name, value in summary:
+                print(f"{field_name}: {value}")
+        else:
+            print("Report generation failed - no metrics collected")
+            return []
+
+        return [ReportFile(name="mock_report", contents=summary)]
